pyproject.toml
[project]
name = "interpretability-bootcamp"
version = "0.1.0"
description = "AI Interpretability Bootcamp - Vector Institute"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argparse>=1.4.0",
    "captum>=0.7.0",
    "dice-ml>=0.9",
    "dnamite>=1.0.5",
    "gosdt>=1.0.1",
    "graphviz>=0.20.3",
    "imblearn>=0.0",
    "imgaug>=0.4.0",
    "interpret>=0.6.9",
    "jupyter>=1.1.1",
    "kaleido==0.2.1",
    "lifelines>=0.30.0",
    "lime>=0.2.0.1",
    "matplotlib>=3.10.0",
    "nodegam>=0.3.0",
    "notebook>=7.3.2",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.0",
    "pyale>=1.2.0",
    "pygam>=0.9.1",
    "python-box[all]~=7.0",
    "pytorch-tabnet>=4.1.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "shap>=0.46.0",
    "tableone>=0.9.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "torchxrayvision>=1.3.2",
    "tqdm>=4.67.1",
    "transformers>=4.48.2",
    "ultraimport>=0.0.7",
    "xgboost>=2.1.3",
    "xlrd>=2.0.1",
]

[dependency-groups]
dev = [
  "ipykernel>=6.29.5",
  "ipython>=8.36.0",
  "ipywidgets>=8.1.7",
  "jupyter>=1.1.1",
  "jupyterlab>=4.4.2",
  "kaggle>=1.7.4.5",
  "codecov>=2.1.13",
  "mypy>=1.14.1",
  "nbqa>=1.9.1",
  "pip-audit>=2.7.3",
  "pre-commit>=4.1.0",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.2",
  "pytest-cov>=6.0.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.9.2",
]

[tool.setuptools]
packages = ["data", "interp", "implementations","scripts", "utils"]
